+++ NEWS.md
@@ -1,3 +1,8 @@
+# clarifai .3
+
+* Support for color, languages, usage 
+* More tests
+
 # clarifai .2
 
 * Bug fix: cases where fewer than 20 tags are returned now handled nicely